// smart-contracts/typechain-types/TipSplitter.ts
export interface TipSplitterInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "admin"
      | "feeAddress"
      | "feeBasisPoints"
      | "setAdmin"
      | "setFeeAddress"
      | "setFeeBasisPoints"
      | "tip"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic:
      | "AdminUpdated"
      | "FeeAddressUpdated"
      | "FeePercentUpdated"
      | "TipSent"
  ): EventFragment;

  encodeFunctionData(functionFragment: "admin", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "feeAddress",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "feeBasisPoints",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "setAdmin",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "setFeeAddress",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "setFeeBasisPoints",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(functionFragment: "tip", values: [AddressLike]): string;

  decodeFunctionResult(functionFragment: "admin", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "feeAddress", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "feeBasisPoints",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "setAdmin", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "setFeeAddress",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setFeeBasisPoints",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "tip", data: BytesLike): Result;
}

export namespace AdminUpdatedEvent {
  export type InputTuple = [oldAdmin: AddressLike, newAdmin: AddressLike];
  export type OutputTuple = [oldAdmin: string, newAdmin: string];
  export interface OutputObject {
    oldAdmin: string;
    newAdmin: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace FeeAddressUpdatedEvent {
  export type InputTuple = [oldAddress: AddressLike, newAddress: AddressLike];
  export type OutputTuple = [oldAddress: string, newAddress: string];
  export interface OutputObject {
    oldAddress: string;
    newAddress: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace FeePercentUpdatedEvent {
  export type InputTuple = [
    oldBasisPoints: BigNumberish,
    newBasisPoints: BigNumberish
  ];
  export type OutputTuple = [oldBasisPoints: bigint, newBasisPoints: bigint];
  export interface OutputObject {
    oldBasisPoints: bigint;
    newBasisPoints: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace TipSentEvent {
  export type InputTuple = [
    sender: AddressLike,
    recipient: AddressLike,
    amount: BigNumberish,
    fee: BigNumberish
  ];
  export type OutputTuple = [
    sender: string,
    recipient: string,
    amount: bigint,
    fee: bigint
  ];
  export interface OutputObject {
    sender: string;
    recipient: string;
    amount: bigint;
    fee: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface TipSplitter extends BaseContract {
  connect(runner?: ContractRunner | null): TipSplitter;
  waitForDeployment(): Promise<this>;

  interface: TipSplitterInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  admin: TypedContractMethod<[], [string], "view">;

  feeAddress: TypedContractMethod<[], [string], "view">;

  feeBasisPoints: TypedContractMethod<[], [bigint], "view">;

  setAdmin: TypedContractMethod<[_newAdmin: AddressLike], [void], "nonpayable">;

  setFeeAddress: TypedContractMethod<
    [_newFeeAddress: AddressLike],
    [void],
    "nonpayable"
  >;

  setFeeBasisPoints: TypedContractMethod<
    [_newBasisPoints: BigNumberish],
    [void],
    "nonpayable"
  >;

  tip: TypedContractMethod<[recipient: AddressLike], [void], "payable">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "admin"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "feeAddress"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "feeBasisPoints"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "setAdmin"
  ): TypedContractMethod<[_newAdmin: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "setFeeAddress"
  ): TypedContractMethod<[_newFeeAddress: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "setFeeBasisPoints"
  ): TypedContractMethod<[_newBasisPoints: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "tip"
  ): TypedContractMethod<[recipient: AddressLike], [void], "payable">;

  getEvent(
    key: "AdminUpdated"
  ): TypedContractEvent<
    AdminUpdatedEvent.InputTuple,
    AdminUpdatedEvent.OutputTuple,
    AdminUpdatedEvent.OutputObject
  >;
  getEvent(
    key: "FeeAddressUpdated"
  ): TypedContractEvent<
    FeeAddressUpdatedEvent.InputTuple,
    FeeAddressUpdatedEvent.OutputTuple,
    FeeAddressUpdatedEvent.OutputObject
  >;
  getEvent(
    key: "FeePercentUpdated"
  ): TypedContractEvent<
    FeePercentUpdatedEvent.InputTuple,
    FeePercentUpdatedEvent.OutputTuple,
    FeePercentUpdatedEvent.OutputObject
  >;
  getEvent(
    key: "TipSent"
  ): TypedContractEvent<
    TipSentEvent.InputTuple,
    TipSentEvent.OutputTuple,
    TipSentEvent.OutputObject
  >;

  filters: {
    "AdminUpdated(address,address)": TypedContractEvent<
      AdminUpdatedEvent.InputTuple,
      AdminUpdatedEvent.OutputTuple,
      AdminUpdatedEvent.OutputObject
    >;
    AdminUpdated: TypedContractEvent<
      AdminUpdatedEvent.InputTuple,
      AdminUpdatedEvent.OutputTuple,
      AdminUpdatedEvent.OutputObject
    >;

    "FeeAddressUpdated(address,address)": TypedContractEvent<
      FeeAddressUpdatedEvent.InputTuple,
      FeeAddressUpdatedEvent.OutputTuple,
      FeeAddressUpdatedEvent.OutputObject
    >;
    FeeAddressUpdated: TypedContractEvent<
      FeeAddressUpdatedEvent.InputTuple,
      FeeAddressUpdatedEvent.OutputTuple,
      FeeAddressUpdatedEvent.OutputObject
    >;

    "FeePercentUpdated(uint256,uint256)": TypedContractEvent<
      FeePercentUpdatedEvent.InputTuple,
      FeePercentUpdatedEvent.OutputTuple,
      FeePercentUpdatedEvent.OutputObject
    >;
    FeePercentUpdated: TypedContractEvent<
      FeePercentUpdatedEvent.InputTuple,
      FeePercentUpdatedEvent.OutputTuple,
      FeePercentUpdatedEvent.OutputObject
    >;

    "TipSent(address,address,uint256,uint256)": TypedContractEvent<
      TipSentEvent.InputTuple,
      TipSentEvent.OutputTuple,
      TipSentEvent.OutputObject
    >;
    TipSent: TypedContractEvent<
      TipSentEvent.InputTuple,
      TipSentEvent.OutputTuple,
      TipSentEvent.OutputObject
    >;
  };
}